Change how we calculate GUID
Previously, we used uuid_generate_time which used the MAC address of the machine as part of the calculation. This is leading to problems when we run on virtual machines as many of the machines share the same dummy MAC address. This has lead to noticeable GUID collisions in production. By switching to uuid_generate_random, we hope to decrease collisions.