Configure the PluginManager not to throw if no cache file
When building a release, the construction of the final edmplugcache file does not happen until the very end of the build. This is after running edmWriteConfigs which can cause the cfi generation to fail.
For now, default user plugins must be linked in with the module using them as a default.