Callback, add clone, store in pair not vector
Add clone method to Callback instead of adding accessors
to internal state for the purpose of copying. In addition,
when building a Callback, optimize by using the fact that
we know the order they are made and store them in a pair
in the factory along with an IOV index instead of holding
all of them in a vector. And also a little cleanup in the
files this touches.