Add optional and required abilities to python parameters
In a PSet one can now associate a label and a type without specifying a default value.
- required: if the value is never set, an exception will be thrown at conversion to C++ time
- optional: nothing happens if the value is never set

This ability is now used to set values for the Process options, maxEvents and maxLuminosityBlock top level PSets

The new options are injected by fillDescriptions.

If more than one type is allowed for a given label, the .allowed mechanism can be used with optional or required.

If any label would be allowed, used the new PSet label 'allowAnyLabel_' and associate a required or optional parameter.