Simplified setting consumes for the EventSetup
The ESGetToken now has a template constructor that expects an Adaptor. This allows the ESGetToken's constructor to pass the two template arguments to the Adaptor so they do not have to be explicitly declared again. The ESProducer, EDConsumerBase, ConsumesCollector and ESConsumesCollector all have helper member functions which are used to generate the appropriate Adaptor.