Prepare runToCompletion() for concurrent runs
Change exceptionMessageRuns_ to atomic<bool> along
with side effects of that. This will be necessary
for concurrent runs and is similar to what we currently
do for lumis.

Plus deletion of some obsolete unused lines of code,
(including an unused set of braces that changes indentation
on a number of lines of code that are otherwise unchanged)